Rearranged code a bit.
