+++ src/CMakeLists.txt
@@ -7,6 +7,7 @@ add_library(rdycore
   rdymesh.c
   rdyrun.c
   rdysetup.c
+  write_output.c
 )
 add_dependencies(rdycore yaml)
